Allow markers to be draggable immediately
When adding new markers to the map, it makes sense to allow the marker to be draggable after it has been places (the user might want to give a more precise location). I'm aware that 0.2 branch has a new "edit mode" to handle this feature, but still I think it's a natural option to be present in a plugin like this.