Events draw:editstart and draw:editstop added
Two additional events are being fired upon activation/deactivation of the
edit mode, similarly to `draw:drawstart` and `draw:drawstop`. I.e. when a
user clicks on the edit or remove button, `draw:editstart` is being fired,
and when the user presses "save", `draw:editstop` is being fired.

With this it becomes possible to e.g. turn off a layer during editing and
switch it back on when editing is finished.