Added workaround for drag lock on map after polygon intersection
There was a bug introduced in Leaflet 1.0.0-rc2 where the map would lock in
a "dragging" state after editing a polygon with an invalid point (when using
allowIntersection=false).  This seemed to be caused by the marker getting
removed from the Poly object while still in the 'dragging' state, and the
Draggable event was never released causing the map to hang.  This was evident
from the 'dragging' mouse cursor on the map, and the 'leaflet-dragging' class
assigned to the <body>.

My workaround is to directly call the private API method to release 'drag':

    marker.dragging._draggable._onUp(e);

There is a related ticket #4484 that has a similar issue where 'drag'
listeners are lost when the Marker is modified (specifically, a setIcon()
call), so I believe this to be an upstream issue.  This issue does not
occur in the 0.7.x branch, so I have added some rudimentary version checks
to avoid running this code on that version.

Unrelated to the 'drag' bug, I reordered some of the lines in that function
to ensure the tooltip error message is actually rendered properly.
The tooltip contents were likely rerendered as part of the onMarkerClick()
call.