[L.Draw.Tooltip] Fix initial position
What:
When clicking on a Toolbar button, the first appearance of the
mouse-following tooltip was in the left corner of the map, instead of the
mouse cursor.

Why:
At initialization time the tooltip lacked the cursor position and
thus, couldn't be properly positioned.

How:
We can't fix easily the unknown cursor position at init time, but we can
easily set a clean workaround.

Instead of trying to show the tooltip when we lack the cursor position, we
now hide it until the cursor position is known. This position is indeed
known starting from the first mousemove event emitted.

So now when a toolbar button is clicked, the tooltip isn't displayed, but
as soon as the user start moving the cursor, the tooltip is displayed
because the cursor position is provided to the tooltip.